requirements.txt
openai
python-dotenv
matplotlib
pandas
